dry delete and update command with an abstract common command
