+++ src/Command/BlacklistDeleteCommand.php
@@ -17,7 +17,7 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Rollerworks\Bundle\PasswordStrengthBundle\Blacklist\SqliteProvider;
 
-class BlacklistDeleteCommand extends BlacklistCommand
+class BlacklistDeleteCommand extends BlacklistCommonCommand
 {
     const MESSAGE = '<info>Successfully removed %d password(s) from your blacklist database.</info>';
 
@@ -34,88 +34,8 @@ class BlacklistDeleteCommand extends BlacklistCommand
         ;
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function attemptAction(SqliteProvider $service, $password)
     {
-        if (!$input->getArgument('passwords') && !$input->getOption('file')) {
-            $output->writeln('<error>No passwords or file-option given.</error>');
-
-            return;
-        }
-
-        /** @var \Rollerworks\Bundle\PasswordStrengthBundle\Blacklist\SqliteProvider $service */
-        $service = $this->getContainer()->get('rollerworks_password_strength.blacklist.provider.sqlite');
-
-        if ($input->getOption('file')) {
-            $file = realpath($input->getOption('file'));
-
-            if (!file_exists($file)) {
-                $output->writeln('<error>Unable to read passwords list. No such file: ' . $input->getOption('file') . '</error>');
-
-                return;
-            }
-
-            if (!is_readable($file)) {
-                $output->writeln('<error>Unable to read passwords list. Access denied: ' . $input->getOption('file') . '</error>');
-
-                return;
-            }
-
-            if (filesize($file) == 0) {
-                $output->writeln('<comment>Passwords list seems empty, are you sure this is the correct file?</comment>');
-
-                return;
-            }
-
-            $count = $this->doFromFile($service, $file);
-        } else {
-            $count = $this->doFromArray($service, (array) $input->getArgument('passwords'));
-        }
-
-        $output->writeln(sprintf(self::MESSAGE, $count));
-    }
-
-    /**
-     * @param SqliteProvider $service
-     * @param string         $filename
-     *
-     * @return integer
-     */
-    protected function doFromFile(SqliteProvider $service, $filename)
-    {
-        $file = new \SplFileObject($filename, 'r');
-        $count = 0;
-
-        foreach ($file as $password) {
-            $password = trim($password, "\n\r");
-            if ($service->isBlacklisted($password) && true === $service->delete($password)) {
-                $count++;
-            }
-        }
-
-        // close file object
-        $file = null;
-
-        return $count;
-    }
-
-    /**
-     * @param SqliteProvider $service
-     * @param array          $passwords
-     *
-     * @return integer
-     */
-    protected function doFromArray(SqliteProvider $service, array $passwords)
-    {
-        $count = 0;
-        foreach ($passwords as $password) {
-            if ($service->isBlacklisted($password) && true === $service->delete($password)) {
-                $count++;
-            }
-        }
-
-        return $count;
+        return $service->isBlacklisted($password) && true === $service->delete($password);
     }
 }
